+++ packages/terra-action-footer/CHANGELOG.md
@@ -3,12 +3,21 @@ ChangeLog
 
 Unreleased
 ----------
-### Changed
-* Import for package version from package.json added to doc page
 
-Unreleased
+2.22.0 - (September 6, 2019)
+------------------
+### Added
+* Props table markdown file import added to doc site page
+
+2.21.0 - (September 3, 2019)
 ------------------
 ### Changed
+* Minor dependency version bump
+
+2.20.0 - (August 21, 2019)
+------------------
+### Changed
+* Import for package version from package.json added to doc page
 * Cleaned up imports in examples and test files
 
 2.19.0 - (August 14, 2019)