+++ packages/terra-action-header/CHANGELOG.md
@@ -3,12 +3,21 @@ ChangeLog
 
 Unreleased
 ----------
-### Changed
-* Import for package version from package.json added to doc page
 
-Unreleased
+2.28.0 - (September 6, 2019)
+------------------
+### Added
+* Props table markdown file import added to doc site page
+
+2.27.0 - (September 3, 2019)
 ------------------
 ### Changed
+* Minor dependency version bump
+
+2.26.0 - (August 21, 2019)
+------------------
+### Changed
+* Import for package version from package.json added to doc page
 * Cleaned up imports in examples and test files
 
 2.25.0 - (August 14, 2019)