+++ packages/terra-alert/CHANGELOG.md
@@ -1,14 +1,24 @@
-Changelog
+ChangeLog
 =========
 
 Unreleased
 ----------
-### Changed
-* Import for package version from package.json added to doc page
 
-Unreleased
+4.11.0 - (September 6, 2019)
+------------------
+### Added
+* Props table markdown file import added to doc site page
+
+4.10.0 - (September 3, 2019)
 ------------------
 ### Changed
+* Minor dependency version bump
+
+4.9.0 - (August 21, 2019)
+------------------
+### Changed
+* Import for package version from package.json added to doc page
+* Switch from Object.assign to Object spread syntax.
 * Cleaned up imports in examples and test files
 
 4.8.0 - (August 14, 2019)