+++ packages/terra-card/CHANGELOG.md
@@ -3,12 +3,19 @@ ChangeLog
 
 Unreleased
 ----------
-### Changed
-* Import for package version from package.json added to doc page
 
-Unreleased
+3.18.0 - (September 6, 2019)
+------------------
+### Added
+* Props table markdown file import added to doc site page
+
+### Removed
+* Removed `inherit` fallback for `--terra-card-color` CSS custom property
+
+3.17.0 - (August 21, 2019)
 ------------------
 ### Changed
+* Import for package version from package.json added to doc page
 * Cleaned up imports in examples and test files
 
 3.16.0 - (August 14, 2019)