+++ packages/terra-dropdown-button/CHANGELOG.md
@@ -3,12 +3,23 @@ ChangeLog
 
 Unreleased
 ----------
-### Changed
-* Import for package version from package.json added to doc page
+### Fixed
+* Reset default margin in Safari.
 
-Unreleased
+1.4.0 - (September 6, 2019)
+------------------
+### Added
+* Props table markdown file import added to doc site page
+
+1.3.0 - (September 3, 2019)
 ------------------
 ### Changed
+* Updated the 'keycode-js' version from '^1.0.4' to '^2.0.1'.
+
+1.2.0 - (August 21, 2019)
+------------------
+### Changed
+* Import for package version from package.json added to doc page
 * Cleaned up imports in examples and test files
 
 1.1.0 - (August 14, 2019)