+++ packages/terra-list/CHANGELOG.md
@@ -1,14 +1,23 @@
-Changelog
+ChangeLog
 =========
 
 Unreleased
 ----------
-### Changed
-* Import for package version from package.json added to doc page
 
-Unreleased
+4.18.0 - (September 6, 2019)
+------------------
+### Added
+* Props table markdown file import added to doc site page
+
+4.17.0 - (September 3, 2019)
 ------------------
 ### Changed
+* Updated the 'keycode-js' version from '^1.0.4' to '^2.0.1'.
+
+4.16.0 - (August 21, 2019)
+------------------
+### Changed
+* Import for package version from package.json added to doc page
 * Cleaned up imports in examples and test files
 
 4.15.0 - (August 14, 2019)