+++ packages/terra-overlay/CHANGELOG.md
@@ -1,14 +1,28 @@
-Changelog
+ChangeLog
 =========
 
 Unreleased
 ----------
-### Changed
-* Import for package version from package.json added to doc page
 
-Unreleased
+3.28.0 - (September 6, 2019)
+------------------
+### Added
+* Props table markdown file import added to doc site page
+
+### Changed
+* Moved running the wicg-inert script to when 'terra-overlay' is mounted to resolve issues in IE 10
+* Update wcig-inert path to import directly the ES5 version of the package
+
+3.27.0 - (September 3, 2019)
 ------------------
 ### Changed
+* Updated the 'keycode-js' version from '^1.0.4' to '^2.0.1'.
+
+3.26.0 - (August 21, 2019)
+------------------
+### Changed
+* Import for package version from package.json added to doc page
+* Update bracket positioning lint
 * Cleaned up imports in examples and test files
 
 3.25.0 - (August 14, 2019)