+++ packages/terra-responsive-element/CHANGELOG.md
@@ -4,7 +4,13 @@ Changelog
 Unreleased
 ----------
 
-Unreleased
+5.8.0 - (September 6, 2019)
+------------------
+### Added
+* Props table markdown file import added to doc site page
+* Added package.json import to doc page
+
+5.7.0 - (August 21, 2019)
 ------------------
 ### Changed
 * Cleaned up imports in examples and test files