+++ packages/terra-search-field/CHANGELOG.md
@@ -1,14 +1,25 @@
-Changelog
+ChangeLog
 =========
 
 Unreleased
 ----------
-### Changed
-* Import for package version from package.json added to doc page
 
-Unreleased
+3.27.0 - (September 6, 2019)
+------------------
+### Added
+* Props table markdown file import added to doc site page
+
+3.26.0 - (September 3, 2019)
 ------------------
 ### Changed
+* updated source and jest files for changes in form-input
+* Updated the 'keycode-js' version from '^1.0.4' to '^2.0.1'.
+
+3.25.0 - (August 21, 2019)
+------------------
+### Changed
+* Import for package version from package.json added to doc page
+* Switch from Object.assign to Object spread syntax.
 * Cleaned up imports in examples and test files
 
 3.24.0 - (August 14, 2019)