+++ packages/terra-show-hide/CHANGELOG.md
@@ -3,13 +3,23 @@ ChangeLog
 
 Unreleased
 ----------
-### Changed
-* Import for package version from package.json added to doc page
 
-Unreleased
+2.22.0 - (September 6, 2019)
+------------------
+### Added
+* Props table markdown file import added to doc site page
+
+2.21.0 - (September 3, 2019)
 ------------------
 ### Changed
+* Updated the 'keycode-js' version from '^1.0.4' to '^2.0.1'.
+
+2.20.0 - (August 21, 2019)
+--------------------------
+### Changed
 * Cleaned up imports in examples and test files
+* Import for package version from package.json added to doc page
+* Added comments for ESlint v6
 
 2.19.0 - (August 14, 2019)
 ------------------